feat: implemented Like logic. Endpoints, service and DTO

+++ src/main/java/com/ms/music_service/repository/LikeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 public interface LikeRepository extends JpaRepository<UserLike, Long> {
-    boolean existsByUserIdAndMusicId(UUID userId, Integer musicId);
+    boolean existsByUserIdAndMusicId(UUID userId, Long musicId);
 }
